fix: move UCSUR utils to shared file for proper importing
